requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain-core
langchain-google-genai
langchain-chroma
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
unstructured[local-inference,ocr]
pytesseract 
Pillow
python-docx 
pdf2image 
protobuf==4.25.3